Rebase, move DirNamesMixin to accessor.py
