+++ pandas/core/strings.py
@@ -1921,19 +1921,3 @@ class StringMethods(NoNewAttributesMixin):
                            "MultiIndex")
                 raise AttributeError(message)
         return cls(data)
-
-
-class StringAccessorMixin(object):
-    """ Mixin to add a `.str` acessor to the class."""
-
-    str = AccessorProperty(StringMethods)
-
-    def _dir_additions(self):
-        return set()
-
-    def _dir_deletions(self):
-        try:
-            getattr(self, 'str')
-        except AttributeError:
-            return set(['str'])
-        return set()
